implementación de los pedido y sus tablas

+++ app/Http/Controllers/AdminController.php
@@ -17,6 +17,12 @@ use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\AddNewImageRequest;
 use App\Http\Requests\editedCarouselContentRequest;
 use App\Http\Requests\SaveProductCategoryRequest;
+use App\SubCategory;
+use App\Http\Requests\SaveSubCategoryRequest;
+use App\Order;
+use App\Unit;
+use App\Http\Requests\AddUnitRequest;
+
 class AdminController extends Controller
 {
     //
@@ -39,8 +45,9 @@ class AdminController extends Controller
     {
         if($request->user()->autorize(1))
         {
+            $units = Unit::all();
             $categories = Category::all();
-            return view('admin.create-product',compact('categories'));
+            return view('admin.create-product',compact(['categories','units']));
         } 
         return back()->with('info','Usted no esta autorizado');
     }
@@ -58,11 +65,14 @@ class AdminController extends Controller
             $product = new Product();
             $product->description =$request->input('description');
             $product->price =$request->input('price');
+            $product->wholesale_price =$request->input('wholesale_price');
+            $product->quantity_wholesale_price =$request->input('quantity_wholesale_price');
             $product->stock =$request->input('stock');
             $product->image =$path.'/'.$time.$image->getClientOriginalName();
             $product->code =$request->input('code');
             $product->long_description =$request->input('long_description');
             $product->id_categories =$request->input('id_categories');
+            $product->id_units =$request->input('id_units');
             $product->save();
             return redirect()->route('panel.admin')->with('info','Producto agregar exitosamente');
         } 
@@ -85,9 +95,10 @@ class AdminController extends Controller
     {
         if($request->user()->autorize([1,3]))
         {
+            $units = Unit::all();
             $product = Product::findOrFail($id);
             $categories = Category::all();
-           return view('admin.edit-product',compact(['product','categories']));
+           return view('admin.edit-product',compact(['product','categories','units']));
         }
         return back()->with('info','Usted no esta autorizado');
     }
@@ -99,10 +110,13 @@ class AdminController extends Controller
             $product = Product::findOrFail($id);
             $product->description =$request->input('description');
             $product->price =$request->input('price');
+            $product->wholesale_price =$request->input('wholesale_price');
+            $product->quantity_wholesale_price =$request->input('quantity_wholesale_price');
             $product->stock =$request->input('stock');
             $product->code =$request->input('code');
             $product->long_description =$request->input('long_description');
             $product->id_categories =$request->input('id_categories');
+            $product->id_units =$request->input('id_units');
             if($request->hasFile('image') !=null)
             {
                 $oldImage = $product->image;
@@ -110,7 +124,7 @@ class AdminController extends Controller
                 $path = 'img/categories/'.$request->input('id_categories');
                 $time = time();
                 $image->move($path,$time.$image->getClientOriginalName()); //save image 
-                Storage::disk('public')->delete($oldImage); 
+                Storage::disk('public')->delete($oldImage); //delete old image
                 $product->image= $path.'/'.$time.$image->getClientOriginalName();
             }
             $product->save();
@@ -152,13 +166,29 @@ class AdminController extends Controller
         return back()->with('info','Usted no esta autorizado');
     }
 
+    public function userMovements(Request $request,$id)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+            $user = User::findOrFail($id);
+            $orders = Order::where('id_user',$id)->get();
+            return view('admin.user-movements',compact(['user','orders']));
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
+
     public function deleteUser(Request $request, $id)
     {
         if($request->user()->autorize(1))
         {
-            $user = User::findOrFail($id);
-            $user->delete();
-            return redirect()->route('admin.users')->with('info','Usuario eleminado con éxito');
+           $user = User::findOrFail($id);
+           $orders = $user->Orders->where('status',Order::PENDING)->first();
+           if(is_null($orders))
+           {
+                $user->delete();
+                return redirect()->route('admin.users')->with('info','Usuario eleminado con éxito');
+           }
+           return back()->with('info',$user->name.' '.$user->last_name.' tiene pedidos por entregar.');
         }
         return back()->with('info','Usted no esta autorizado');
     }
@@ -228,7 +258,7 @@ class AdminController extends Controller
                 Storage::disk('public')->delete($oldImage);
             }
             $carousel->save();
-           return redirect()->route('admin.carousel.view')->with('info','Contenido del carrusel altualizado con éxito');
+           return redirect()->route('admin.carousel.view')->with('info','Contenido del carrusel actualizado con éxito');
         }
         return back()->with('info','Usted no esta autorizado');
     }
@@ -250,7 +280,8 @@ class AdminController extends Controller
         if($request->user()->autorize([1,3]))
         {
             $categories = Category::all();
-            return view('admin.categories-view',compact('categories'));
+            $subCategories = SubCategory::all();
+            return view('admin.categories-view',compact(['categories','subCategories']));
         }
         return back()->with('info','Usted no esta autorizado');
     }
@@ -272,7 +303,8 @@ class AdminController extends Controller
         if($request->user()->autorize([1,3]))
         {
             $category = Category::findOrFail($id);
-            return view('admin.category-edit',compact('category'));
+            $subCategories = SubCategory::all();
+            return view('admin.category-edit',compact(['category','subCategories']));
         }
         return back()->with('info','Usted no esta autorizado');
     }
@@ -283,6 +315,7 @@ class AdminController extends Controller
         {
             $category = Category::findOrFail($id);
             $category->category=$request->input('category');
+            $category->id_SubCategory=$request->input('id_SubCategory');
             $category->save();
             return redirect()->route('admin.categories.view')->with('info','Se actualizó una categoría');
         }
@@ -294,10 +327,143 @@ class AdminController extends Controller
         if($request->user()->autorize(1))
         {
             $category = Category::findOrFail($id);
+            $products = $category->Product->where('id_categories',$id)->first();
+            if($products != null)
+            {
+                return back()->with('info','No puedes eliminar esta categoría, ya que está en uso.');
+            }
             $category->delete();
             return redirect()->route('admin.categories.view')->with('info','Se eliminó una categoría');
         }
         return back()->with('info','Usted no esta autorizado');
     }
+
+    public function addSubcategory(SaveSubCategoryRequest $request)
+    {
+        if($request->user()->autorize(1))
+        {
+            $subCategory = new SubCategory();
+            $subCategory->fill($request->all());
+            $subCategory->save();
+            return back()->with('info','Se agregó una subcategoría');
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
     
+    public function editSubcategory(Request $request, $id)
+    {
+        if($request->user()->autorize(1))
+        {
+            $subCategory = SubCategory::findOrFail($id);
+            return view('admin.subCategory-edit',compact('subCategory'));
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
+
+    public function editedSubcategory(SaveSubCategoryRequest $request, $id)
+    {
+        if($request->user()->autorize(1))
+        {
+            $subCategory = SubCategory::findOrFail($id);
+            $subCategory->description = $request->input('description');
+            $subCategory->save();
+            return redirect()->route('admin.categories.view')->with('info','Se ha actualizado una subcategoría');
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
+
+    public function deleteSubcategory(Request $request,$id)
+    {
+        if($request->user()->autorize(1))
+        {
+            $category = Category::where('id_SubCategory',$id)->first();
+            if(is_null($category))
+            {
+                $subCategory = SubCategory::findOrFail($id);
+                $subCategory->delete();
+                return back()->with('info','Se ha eliminado una subcategoría.');     
+            }
+            return back()->with('info','No es posible eliminar este  registro, porque está en uso.');
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
+
+    public function unitsView(Request $request)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+            $units = Unit::all();
+            return view('admin.units',compact('units'));
+        }
+        return back()->with('info','Usted no esta autorizado');
+    }
+
+    public function addUnit(AddUnitRequest $request)
+    {
+        if($request->user()->autorize(1))
+        {
+            $unit = new Unit();
+            $unit->fill($request->all());
+            $unit->save();
+            return back()->with('info','Unidad agregada con éxito.');
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
+
+    public function editUnit(Request $request,$id)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+            $unit = Unit::findOrFail($id);
+            return view('admin.unit-edit',compact('unit'));
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
+
+    public function editedUnit(AddUnitRequest $request,$id)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+            $unit = Unit::findOrFail($id);
+            $unit->description = $request->input('description');
+            $unit->save();
+            return redirect()->route('admin.units.view')->with('info','Unidad actualizada con éxito.');
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
+
+    public function deleteUnit(Request $request,$id)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+          $unit = Unit::findOrFail($id);
+          $products = $unit->Products->where('id_units',$id)->first();
+          if($products != null)
+          {
+            return back()->with('info','No puede eliminar este campo, ya que está en uso en los productos.');
+          }
+          $unit->delete();
+           return back()->with('info','Unidad eliminada con éxito.');
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
+
+    public function ordersView(Request $request)
+    {
+        if($request->user()->autorize([1,3]))
+        {
+            $orders = Order::where('status',Order::PENDING)->get();
+            return view('admin.orders-view',compact('orders'));
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
+
+    public function orderSended(Request $request,$id)
+    {
+        if($request->user()->autorize([1,3,4]))
+        {
+            return $id;
+        }
+        return back()->with('info','Usted no esta autorizado.');
+    }
 }
